requirements.txt
openai
python-dotenv
yfinance
pandas
requests
